Adds a version of PubGraph that fully creates the neo4j graph.
